+++ intelmq/bots/experts/do_portal/expert.py
@@ -1,9 +1,15 @@
 # -*- coding: utf-8 -*-
+"""
+As the frontend reverse-proxies the (backend) API
+a "502 Bad Gateway" status code is treated the same as a timeout,
+i.e. will be retried instead of a fail.
+"""
 try:
     import requests
 except ImportError:
     requests = None
 
+import intelmq.lib.utils as utils
 from intelmq.lib.bot import Bot
 
 
@@ -22,6 +28,13 @@ class DoPortalExpertBot(Bot):
         })
         self.mode = self.parameters.mode
 
+        self.session = utils.create_request_session_from_bot(self)
+        retries = requests.urllib3.Retry.from_int(self.http_timeout_max_tries)
+        retries.status_forcelist = [502]
+        adapter = requests.adapters.HTTPAdapter(max_retries=retries)
+        self.session.mount('http://', adapter)
+        self.session.mount('https://', adapter)
+
     def process(self):
         event = self.receive_message()
         if "source.ip" not in event:
@@ -29,24 +42,7 @@ class DoPortalExpertBot(Bot):
             self.acknowledge_message()
             return
 
-        timeoutretries = 0
-        req = None
-
-        while timeoutretries < self.http_timeout_max_tries and req is None:
-            try:
-                req = requests.get(self.url % event['source.ip'],
-                                   headers=self.http_header,
-                                   auth=self.auth,
-                                   proxies=self.proxy,
-                                   verify=self.http_verify_cert,
-                                   cert=self.ssl_client_cert,
-                                   timeout=self.http_timeout_sec)
-            except requests.exceptions.Timeout:
-                timeoutretries += 1
-
-        if req is None and timeoutretries >= self.http_timeout_max_tries:
-            raise ValueError("Request timed out %i times in a row."
-                             "" % timeoutretries)
+        req = self.session.get(self.url % event['source.ip'])
 
         if req.status_code == 404 and req.json()['message'].startswith("('no such cidr'"):
             result = []