+++ intelmq/bots/experts/ripe/expert.py
@@ -8,6 +8,7 @@ https://github.com/RIPE-NCC/whois/wiki/WHOIS-REST-API-abuse-contact
 import json
 from contextlib import contextmanager
 
+import intelmq.lib.utils as utils
 from intelmq.lib.bot import Bot
 from intelmq.lib.cache import Cache
 
@@ -74,12 +75,8 @@ class RIPEExpertBot(Bot):
         self.__initialize_cache()
 
     def __initialize_http_session(self):
-        self.http_session = requests.Session()
         self.set_request_parameters()
-        self.http_session.proxies.update(self.proxy)
-        self.http_session.headers.update(self.http_header)
-        self.http_session.verify = self.http_verify_cert
-        self.http_session.cert = self.ssl_client_cert
+        self.http_session = utils.create_request_session_from_bot(self)
 
     def __initialize_cache(self):
         cache_host = getattr(self.parameters, 'redis_cache_host')
@@ -137,19 +134,8 @@ class RIPEExpertBot(Bot):
             else:
                 return json.loads(cached_value)
         else:
-            timeoutretries = 0
-            response = None
-
-            while timeoutretries < self.http_timeout_max_tries and response is None:
-                try:
-                    response = self.http_session.get(self.QUERY[type].format(resource),
-                                                     data="", timeout=self.http_timeout_sec)
-                except requests.exceptions.Timeout:
-                    timeoutretries += 1
-
-            if response is None and timeoutretries >= self.http_timeout_max_tries:
-                raise ValueError("Request timed out %i times in a row."
-                                 "" % timeoutretries)
+            response = self.http_session.get(self.QUERY[type].format(resource),
+                                             data="", timeout=self.http_timeout_sec)
 
             if response.status_code != 200:
                 if type == 'db_asn' and response.status_code == 404: